render image(s) to video file; refactor baker/encoder configuration approach; add option to add initial transform values for layers; queue layer events if not initialized;

+++ sample/src/main/java/com/netherpyro/glcv/MainActivity.kt
@@ -38,7 +38,7 @@ class MainActivity : AppCompatActivity() {
         override fun onVideoSizeChanged(width: Int, height: Int, unappliedRotationDegrees: Int,
                                         pixelWidthHeightRatio: Float) {
             transformableList.findTransformable(videoTag)
-                ?.setSize(width * pixelWidthHeightRatio, height * pixelWidthHeightRatio)
+                ?.setSize((width * pixelWidthHeightRatio).toInt(), (height * pixelWidthHeightRatio).toInt())
         }
 
         override fun onRenderedFirstFrame() {}
